requirements.txt
scikit-learn>=1.3.0
librosa>=0.10.1
numpy>=1.25.0
soundfile>=0.12.1
pydub>=0.25.1
joblib>=1.3.0
pytest>=7.4.0
pytest-cov>=4.1.0
openai-whisper>=20231117
num2words>=0.5.13
Flask>=2.3.3
